merge Develop to main

+++ models/Unidad.php
@@ -16,6 +16,14 @@ class Unidad extends Conectar
         $params = [':unid_id' => $unid_id];
         return $this->ejecutarConsulta($sql, $params);
     }
+   public function get_seccion_unidad($sec_id)
+    {
+    $sql = "SELECT * FROM tm_seccion sec
+    INNER JOIN tm_unidad uni ON uni.unid_id = sec.sec_unidad
+    WHERE sec_id = :sec_id";
+        $params = [':sec_id' => $sec_id];
+        return $this->ejecutarConsulta($sql, $params);
+    }
 
     // get_unidad_est según disponibilidad
     public function get_unidad_est($unid_est)
@@ -26,19 +34,16 @@ class Unidad extends Conectar
     }
 
     // add_unidad
-    public function add_unidad($unid_nom, $unid_est, $responsable_rut, $reemplazante_rut)
+    public function add_unidad($unid_nom, $unid_est)
     {
-        if (empty($unid_nom) || empty($unid_est) || empty($responsable_rut) || empty($reemplazante_rut)) {
+        if (empty($unid_nom) || empty($unid_est) ) {
             return ['status' => 'warning', 'message' => 'Todos los campos son obligatorios.'];
         }
 
-        $sql = "INSERT INTO tm_unidad (unid_nom, unid_est, responsable_rut, reemplazante_rut) 
-                VALUES (:unid_nom, :unid_est, :responsable_rut, :reemplazante_rut)";
+        $sql = "INSERT INTO tm_unidad (unid_nom, unid_est) VALUES (:unid_nom, :unid_est)";
         $params = [
             ':unid_nom' => $unid_nom,
             ':unid_est' => $unid_est,
-            ':responsable_rut' => $responsable_rut,
-            ':reemplazante_rut' => $reemplazante_rut
         ];
 
         if ($this->ejecutarAccion($sql, $params)) {
@@ -51,24 +56,20 @@ class Unidad extends Conectar
     }
 
     // update_unidad según id
-    public function update_unidad($unid_id, $unid_nom, $unid_est, $responsable_rut, $reemplazante_rut)
+    public function update_unidad($unid_id, $unid_nom, $unid_est)
     {
-        if (empty($unid_id) || empty($unid_nom) || empty($unid_est) || empty($responsable_rut) || empty($reemplazante_rut)) {
+        if (empty($unid_id) || empty($unid_nom) || empty($unid_est)) {
             return ['status' => 'warning', 'message' => 'Todos los campos son obligatorios.'];
         }
 
         $sql = "UPDATE tm_unidad SET 
                     unid_nom = :unid_nom, 
                     unid_est = :unid_est, 
-                    responsable_rut = :responsable_rut, 
-                    reemplazante_rut = :reemplazante_rut 
                 WHERE unid_id = :unid_id";
         $params = [
             ':unid_id' => $unid_id,
             ':unid_nom' => $unid_nom,
             ':unid_est' => $unid_est,
-            ':responsable_rut' => $responsable_rut,
-            ':reemplazante_rut' => $reemplazante_rut
         ];
 
         if ($this->ejecutarAccion($sql, $params)) {
@@ -83,7 +84,10 @@ class Unidad extends Conectar
     {
         try {
             // Verificar si la unidad está en uso en otras tablas
-            $sql_check_evun = "SELECT * FROM tm_asignado WHERE unid_id = :unid_id";
+            $sql_check_evun = "SELECT * FROM tm_asignado
+                         INNER JOIN tm_seccion ON tm_asignado.sec_id = tm_seccion.sec_id
+                         INNER JOIN tm_unidad ON tm_seccion.sec_unidad = tm_unidad.unid_id
+                         WHERE unid_id = :unid_id";
             $params = [':unid_id' => $unid_id];
             if ($this->ejecutarConsulta($sql_check_evun, $params)) {
                 return ['status' => 'warning', 'message' => 'La unidad no se puede eliminar porque está en uso en otros registros.'];
@@ -93,6 +97,11 @@ class Unidad extends Conectar
             if ($this->ejecutarConsulta($sql_check_usuario, $params)) {
                 return ['status' => 'warning', 'message' => 'La unidad no se puede eliminar porque está en uso en otros registros.'];
             }
+            $sql_check_seccion = "SELECT * FROM tm_seccion WHERE sec_unidad=:unid_id";
+            if ($this->ejecutarConsulta($sql_check_seccion, $params)) {
+                return ['status' => 'warning', 'message' => 'La unidad no se puede eliminar porque está en uso en otros registros.'];
+            }
+
 
             // La unidad no está en uso, proceder a eliminarla
             $sql_delete = "DELETE FROM tm_unidad WHERE unid_id = :unid_id";
